requirements.txt
google-generativeai
Pillow
flask